Move CollectionSizeAttribute to AutoData
